--- backend/src/main/java/com/projectalpha/exception/UserNotVerifiedException.java
@@ -1,16 +1,0 @@
-package com.projectalpha.exception;
-
-public class UserNotVerifiedException extends RuntimeException {
-
-    public UserNotVerifiedException() {
-        super("User is not verified waiting for new OTP authentication.");
-    }
-
-    public UserNotVerifiedException(String message) {
-        super(message);
-    }
-
-    public UserNotVerifiedException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
