+++ backend/src/main/java/com/projectalpha/exception/auth/DuplicateEmailException.java
@@ -1,0 +1,19 @@
+package com.projectalpha.exception.auth;
+
+/**
+ * Exception thrown when attempting to register with an email that is already in use.
+ */
+public class DuplicateEmailException extends RuntimeException {
+    
+    public DuplicateEmailException() {
+        super("Email is already registered");
+    }
+    
+    public DuplicateEmailException(String message) {
+        super(message);
+    }
+    
+    public DuplicateEmailException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
