+++ backend/src/main/java/com/projectalpha/exception/auth/VerificationCodeNotCorrect.java
@@ -1,0 +1,15 @@
+package com.projectalpha.exception.auth;
+
+public class VerificationCodeNotCorrect extends RuntimeException {
+    public VerificationCodeNotCorrect() {
+        super("Verification code not correct");
+    }
+
+    public VerificationCodeNotCorrect(String message) {
+        super(message);
+    }
+
+    public VerificationCodeNotCorrect(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
