+++ backend/src/main/java/com/projectalpha/exception/auth/WrongRoleLoginMethod.java
@@ -1,0 +1,15 @@
+package com.projectalpha.exception.auth;
+
+public class WrongRoleLoginMethod extends RuntimeException {
+    public WrongRoleLoginMethod() {
+        super("Wrong login method called");
+    }
+
+    public WrongRoleLoginMethod(String message) {
+        super(message);
+    }
+
+    public WrongRoleLoginMethod(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
