+++ backend/src/main/java/com/projectalpha/service/user/UserService.java
@@ -1,0 +1,109 @@
+package com.projectalpha.service.user;
+
+/*import com.projectalpha.dto.ListDTO;
+import com.projectalpha.dto.ReviewDTO;
+import com.projectalpha.dto.UserProfileDTO;*/
+import com.projectalpha.dto.user.diner.DinerUserProfile;
+import com.projectalpha.dto.user.owner.OwnerUserProfile;
+import com.projectalpha.repository.user.diner.DinerRepository;
+import com.projectalpha.repository.user.owner.OwnerRepository;
+import com.projectalpha.service.user.diner.DinerService;
+import com.projectalpha.service.user.owner.OwnerService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
+//import java.util.List;
+
+@Service
+public class UserService implements DinerService, OwnerService {
+
+    private final DinerRepository dinerRepository;
+    private final OwnerRepository ownerRepository;
+
+    @Autowired
+    public UserService(DinerRepository dinerRepository, OwnerRepository ownerRepository) {
+        this.dinerRepository = dinerRepository;
+        this.ownerRepository = ownerRepository;
+    }
+
+    @Override
+    public Optional<DinerUserProfile> getDinerProfileByUserId(String userId) {
+        return dinerRepository.findDinerByID(userId);
+    }
+
+    @Override
+    public Optional<OwnerUserProfile> getOwnerProfileByUserId(String userId) {
+        return ownerRepository.findOwnerByID(userId);
+    }
+
+    @Override
+    public void updateProfile(String userId, DinerUserProfile profile) {
+        dinerRepository.updateDinerProfile(userId, profile);
+    }
+
+    @Override
+    public void updateProfile(String userId, OwnerUserProfile profile) {
+        ownerRepository.updateOwnerProfile(userId, profile);
+    }
+
+}
+
+
+// /**
+// * Service interface for user-related operations
+// */
+// old implementations
+// /**
+// */
+// * Update a user's user information
+// *
+// * @param userId The ID of the user
+// * @param profileData The updated user data
+// */
+//void updateUserProfile(String userId, UserProfileDTO profileData);
+
+// /**
+// * Get all lists for a user (including favorites)
+// *
+// * @param userId The ID of the user
+// * @return List of user's lists
+// */
+// List<ListDTO> getUserLists(String userId);
+
+// /**
+// * Create a new list for a user
+// *
+// * @param userId The ID of the user
+// * @param listName The name of the new list
+// * @return The created list
+// */
+// ListDTO createList(String userId, String listName);
+
+// /**
+// * Add a business to a user's list
+// *
+// * @param userId The ID of the user
+// * @param listId The ID of the list
+// * @param businessId The ID of the business to add
+// */
+// void addBusinessToList(String userId, String listId, String businessId);
+
+// /**
+// * Remove a business from a user's list
+// *
+// * @param userId The ID of the user
+// * @param listId The ID of the list
+// * @param businessId The ID of the business to remove
+// */
+// void removeBusinessFromList(String userId, String listId, String businessId);
+
+// /**
+// * Get all reviews written by a user
+// *
+// * @param userId The ID of the user
+// * @return List of user's reviews
+// */
+// List<ReviewDTO> getUserReviews(String userId);
+//
