+++ backend/src/main/java/com/projectalpha/service/user/owner/OwnerService.java
@@ -1,0 +1,23 @@
+package com.projectalpha.service.user.owner;
+
+import com.projectalpha.dto.user.owner.OwnerUserProfile;
+
+import java.util.Optional;
+
+public interface OwnerService {
+    /**
+     * Get a user's user information
+     *
+     * @param userId The ID of the user
+     * @return User user data
+     */
+    Optional<OwnerUserProfile> getOwnerProfileByUserId(String userId);
+
+    /**
+     * Update ownerProfile
+     *
+     * @param userId ownerId
+     * @param profile ownerProfile
+     */
+    void updateProfile(String userId, OwnerUserProfile profile);
+}
