+++ backend/src/main/resources/application.properties
@@ -2,9 +2,12 @@ spring.application.name=projectalpha
 spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
 spring.security.user.name=admin
 spring.security.user.password=1234
+spring.datasource.password = eRnhwuEDKKF8FHvX
+spring.datasource.username = postgres.jfqmvaeeelneeawbvkga
 
 supabase.url=https://jfqmvaeeelneeawbvkga.supabase.co
 supabase.apiKey=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImpmcW12YWVlZWxuZWVhd2J2a2dhIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NDU3MDAwNzgsImV4cCI6MjA2MTI3NjA3OH0.B9R9mUucwZCYXSuZPD3PKGcNUfR58Zb4Efb0kyUe2RM
 supabase.secretKey=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImpmcW12YWVlZWxuZWVhd2J2a2dhIiwicm9sZSI6InNlcnZpY2Vfcm9sZSIsImlhdCI6MTc0NTcwMDA3OCwiZXhwIjoyMDYxMjc2MDc4fQ.WG-0ef4z9yf84U5XfQR6_iMaObMMWZwVjga6u5fNu2w
 logging.level.org.springframework.web.client.RestTemplate=DEBUG
 logging.level.org.apache.http=DEBUG
+
